+++ src/OpenLoco/GameCommands/GameCommands.h
@@ -77,7 +77,7 @@ namespace OpenLoco::GameCommands
         gc_unk_41 = 41,
         gc_unk_42 = 42,
         removeRoadStation = 43,
-        gc_unk_44 = 44,
+        createBuilding = 44,
         removeBuilding = 45,
         renameTown = 46,
         createIndustry = 47,
@@ -596,11 +596,54 @@ namespace OpenLoco::GameCommands
         return doCommand(GameCommand::removeRoadStation, regs) != FAILURE;
     }
 
+    struct BuildingPlacementArgs
+    {
+        BuildingPlacementArgs() = default;
+        explicit BuildingPlacementArgs(const registers regs)
+            : pos(regs.ax, regs.cx, regs.di)
+            , rotation(regs.bh & 0x3)
+            , type(regs.dl)
+            , variation(regs.dh)
+            , colour(regs.edi >> 16)
+            , buildImmediately(regs.bh & 0x80)
+        {
+        }
+
+        Map::Pos3 pos;
+        uint8_t rotation;
+        uint8_t type;
+        uint8_t variation;
+        Colour_t colour;
+        bool buildImmediately = false; // No scaffolding required (editor mode)
+        explicit operator registers() const
+        {
+            registers regs;
+            regs.ax = pos.x;
+            regs.cx = pos.y;
+            regs.edi = pos.z | (colour << 16);
+            regs.dl = type;
+            regs.dh = variation;
+            regs.bh = rotation | (buildImmediately ? 0x80 : 0);
+            return regs;
+        }
+    };
+
+    inline bool do_44(const BuildingPlacementArgs& placementArgs, uint8_t flags)
+    {
+        registers regs = registers(placementArgs);
+        regs.bl = flags;
+        return doCommand(GameCommand::createBuilding, regs) != FAILURE;
+    }
+
     struct BuildingRemovalArgs
     {
         BuildingRemovalArgs() = default;
         explicit BuildingRemovalArgs(const registers regs)
             : pos(regs.ax, regs.cx, regs.di)
+        {
+        }
+        explicit BuildingRemovalArgs(const BuildingPlacementArgs& place)
+            : pos(place.pos)
         {
         }